backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<3.16
django-cors-headers>=4.3.0
